remove single-event API, move resolver to frontend-internal
