+++ cmd/frontend/internal/telemetry/resolvers/BUILD.bazel
@@ -1,10 +1,10 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "resolver",
-    srcs = ["resolver.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/internal/telemetry/resolver",
-    visibility = ["//:__subpackages__"],
+    name = "resolvers",
+    srcs = ["resolvers.go"],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/telemetry/resolvers",
+    visibility = ["//cmd/frontend:__subpackages__"],
     deps = [
         "//cmd/frontend/graphqlbackend",
         "//internal/database",
